analise/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
sqlalchemy
psycopg2-binary 